ui refactoring & cleanups
