+++ CMakeLists.txt
@@ -36,6 +36,7 @@ add_executable(${PROJECT_NAME}.elf
 	src/gui/ime.c
 	src/gui/ui.c
 	src/gui/ui_settings.c
+	src/gui/ui_connect.c
 
 	libgamestream/client.c
 	libgamestream/http.c
