+++ src/gui/ui_connect.h
@@ -1,0 +1,7 @@
+#include <stdbool.h>
+
+void ui_connect_address(char *addr);
+bool ui_connect_connected();
+
+void ui_connect_saved();
+void ui_connect_ip();
