+++ src/gui/ui_settings.h
@@ -1,3 +1,3 @@
 
-int settings_menu();
-void settings_save_config();
+int ui_settings_menu();
+void ui_settings_save_config();
